the feelings component is made
